JBM: Add solid() and empty() members to Tile

+++ simplegame/gamelib_story_screen.cpp
@@ -6,8 +6,8 @@
 namespace GameLib {
 	StoryScreen::StoryScreen() {
 		Context* context = Locator::getContext();
-		screenWidth = context->screenWidth;
-		screenHeight = context->screenHeight;
+		screenWidth = (float)context->screenWidth;
+		screenHeight = (float)context->screenHeight;
 		screenSize = std::min(screenWidth, screenHeight);
 	}
